Create parkcount.py
The parking count.
Input is two digit (x,y)
x = hill number (hill between floor 1 and 2 is 1, 2 and 3 is 2, etc)
y = in/out (1/0) 
example:
11 means a car went from 1->2
10 means a car went from 2->1

note: each index of the array corresponds to each floor. The values are the number of open spots. Index 0 is the total number of cars entering (level 0 tag)