Implemented proper design
created a method to take car of the car count. It works assuming the input is a two digit number, the first digit being the "hill" and the second being "direction." Hill 0 is between 0 and 1 aka the entrance, hill 1 is between floor 1 and 2, hill 2 between 2 and 3, etc.  the direction input is either 1 or 0. 0 being in and 1 being out.